PatchGuard/requirements.txt
# Test dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Production dependencies
PyGithub>=2.0.0
httpx>=0.24.0
redis>=4.5.0
kubernetes>=27.0.0
prometheus-client>=0.17.0
pyjwt>=2.8.0
cryptography>=41.0.0

# LLM providers
google-generativeai>=0.3.0
